fix(6718): MATS submission requires DSQA facility permission
